pass `AppLoadContext` to the server's `handleRequest` function
The `AppLoadContext` instance can optionally be set via the server adapter's `getLoadContext()` function. It is subsequently passed to all actions, loaders as well as the `handleDataRequest` function. It is, however, not passed to the `handleRequest` function.

This PR also passes the load context to the `handleRequest` function so that the context is available everywhere in a consistent manner.